@@ -1,49 +1,16 @@
-// This is a manifest file that'll be compiled into including all the files listed below.
-// Add new JavaScript/Coffee code in separate files in this directory and they'll automatically
-// be included in the compiled file accessible from http://example.com/assets/application.js
+// This is a manifest file that'll be compiled into application.js, which will include all the files
+// listed below.
+//
+// Any JavaScript/Coffee file within this directory, lib/assets/javascripts, vendor/assets/javascripts,
+// or vendor/assets/javascripts of plugins, if any, can be referenced here using a relative path.
+//
 // It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
 // the compiled file.
 //
+// WARNING: THE FIRST BLANK LINE MARKS THE END OF WHAT'S TO BE PROCESSED, ANY BLANK LINE SHOULD
+// GO AFTER THE REQUIRES BELOW.
+//
 //= require jquery
 //= require jquery_ujs
+//= require foundation
 //= require_tree .
-//= require modernizr
-
-function showDaysValue(newValue)
-{
-  if (newValue > 1) {
-  	document.getElementById("daysrange").innerHTML=newValue + ' Days';
-  } else {
-  	document.getElementById("daysrange").innerHTML=newValue + ' Day';    
-  }
-}
-
-function showViewsValue(newValue)
-{
-  if (newValue > 1) {
-  	document.getElementById("viewsrange").innerHTML=newValue + ' Views';
-  } else {
-  	document.getElementById("viewsrange").innerHTML=newValue + ' View';
-  }
-}
-
-
-msg = "Enter the Password to be Shared"
-function prepareTextField(e) {
-  if (e) {
-    if (e.value == msg) {
-      e.value = '';
-    }
-  }
-}
-
-function revertTextField(e) {
-  if (e)
-    if (e.value == '') {
-      e.value = msg;
-    }
-}
-
-function setCopied() {
-	$('#clip_tip').text('copied!');
-}
